feat: horrendous decor issue on save const fixed via recreate DataArrayTexture + upgrade react-three-fiber and three

+++ packages/ui/world/src/components/Ceiling.tsx
@@ -130,7 +130,7 @@ export default function Ceiling() {
     const texArray = w.texCeil;
     const uvDims = attribute("uvDimensions", "vec2");
     const uvOffs = attribute("uvOffsets", "vec2");
-    const uvTexIds = attribute("uvTextureIds", "float");
+    const uvTexIds = attribute("uvTextureIds", "uint");
     const transformedUv = uv().mul(uvDims).add(uvOffs);
     const texNode = texture(texArray.tex, transformedUv);
     texNode.depthNode = instanceIndex.mod(int(texArray.opts.numTextures));
